[WFCORE-4000] Move JdkType from "version" module to "process-controller" module.
