+++ core-feature-pack/src/main/resources/modules/system/layers/base/org/jboss/as/version/main/module.xml
@@ -32,8 +32,6 @@
     </resources>
 
     <dependencies>
-        <module name="org.jboss.logging"/>
-        <module name="org.jboss.logmanager"/>
         <module name="org.jboss.modules"/>
     </dependencies>
 </module>
